+++ java/src/main/java/io/monohull/dto/ImageConfigBundle.java
@@ -42,6 +42,7 @@ public record ImageConfigBundle(
         String dbCommand,
         String hostVolumePath,
         String dbVolumeName,
+        String dbVolumeTarget,
         String workspacePath,
         Integer appHttpPort,
         Integer appHttpsPort,