+++ java/src/main/java/io/monohull/dto/ImageConfigResponse.java
@@ -17,6 +17,7 @@ public record ImageConfigResponse(
     String dbCommand,
     String hostVolumePath,
     String dbVolumeName,
+    String dbVolumeTarget,
     String workspacePath,
     Integer appHttpPort,
     Integer appHttpsPort,