+++ java/src/test/java/io/monohull/service/BundleServiceTest.java
@@ -53,7 +53,7 @@ class BundleServiceTest {
             "acme", "eam", "7.6.1.3",
             "app:1", "db:1", "adm:1",
             "DB2", "maxdb76", null, null,
-            null, null, null,
+            null, null, null, null,
             null, null, null, null, null, null,
             null, null, null, null, null, null);
     }